Load images by id on game cards, organize game catalog in grid layout
