+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {BrowserRouter, Switch, Route} from "react-router-dom";
 import './App.scss';
 import Navbar from "./components/Navbar/Navbar";
-import GameCatalog from "./views/GameCatalog";
+import GameCatalog from "./views/GameCatalog/GameCatalog";
 import GameDetails from "./views/GameDetails/GameDetails";
 import NewTemplate from './views/NewTemplate/NewTemplate';
 import GameTemplates from './views/GameTemplates/GameTemplates';
