+++ src/components/GameCard/GameCard.js
@@ -5,11 +5,34 @@ import './GameCard.scss';
 
 class GameCard extends Component {
 
+    constructor(props) {
+        super(props);
+        this.state = {
+            image: null
+        }
+    }
+
+    componentDidMount() {
+        try {
+            const gameImage = require(`../../games/${this.props.id}/game-image.png`);
+            if (gameImage) {
+                this.setState({
+                    image: gameImage
+                });
+            }
+        } catch (e) {
+            console.warn(`${e.message}: The game of id ${this.props.id} does not have game-image.png`);
+        }
+    }
+
     render() {
         return (
             <Link to={`/games/${this.props.id}`} className="link-wrapper">
                 <div className="game-card">
-                    <div style={{backgroundColor: !!this.props.color ? this.props.backgroundColor : "white"}} className="game-image">Image placeholder</div>
+                    {!!this.state.image ?
+                        <img className="game-image" src={this.state.image} alt="Game"/> :
+                        <div className="game-image" style={{backgroundColor: !!this.props.color ? this.props.color : "#2d2d2d"}}/>
+                    }
                     <div className="game-info">
                         <h3>{!!this.props.name ? this.props.name : "Game Title"}</h3>
                         <p>{!!this.props.description ? this.props.description : ""}</p>
@@ -27,7 +50,7 @@ GameCard.propTypes = {
     color: PropTypes.string,
     name: PropTypes.string.isRequired,
     description: PropTypes.string.isRequired,
-    updated: PropTypes.string.isRequired
+    updated: PropTypes.string.isRequired,
 }
 
 export default GameCard;
