+++ src/components/GameCard/GameCard.scss
@@ -23,7 +23,6 @@
 
   .game-image {
     border-radius: 4px 4px 0 0;
-    background-color: #484848;
     width: 100%;
     height: 160px;
     text-align: center;
